organizing firebase file
