+++ App.js
@@ -6,23 +6,8 @@ import LoginScreen from './screens/LoginScreen'
 import RegisterScreen from './screens/RegisterScreen'
 import CategorySignin from './screens/CategorySignin'
 import FeedScreen from './screens/FeedScreen'
-import * as firebase from 'firebase'
 
-//firebacse key 
-var firebaseConfig = {
-  apiKey: "AIzaSyCN8ZmOhyI2upGCWpPzMZsqmEY_igIvrgc",
-  authDomain: "sparked-bd638.firebaseapp.com",
-  databaseURL: "https://sparked-bd638.firebaseio.com",
-  projectId: "sparked-bd638",
-  storageBucket: "sparked-bd638.appspot.com",
-  messagingSenderId: "480536595337",
-  appId: "1:480536595337:web:945f6e133d77bc19a66f58",
-  measurementId: "G-YGGLN3DBXH"
-};
-
-// Initialize Firebase
-firebase.initializeApp(firebaseConfig);
-//firebase.analytics();
+import firebase from './firebase'
 
 //navivation once logged in 
 const LoginStack = createStackNavigator({
@@ -31,9 +16,10 @@ const LoginStack = createStackNavigator({
 
 //auth navigation
 const AuthStack = createStackNavigator({
-   Login: LoginScreen,
+  Login: LoginScreen,
   Register: RegisterScreen, 
   pickCategory : CategorySignin,
+
 })
 
 //create navigation
