+++ package.json
@@ -9,6 +9,9 @@
   },
   "dependencies": {
     "@react-native-community/masked-view": "0.1.5",
+    "@react-navigation/bottom-tabs": "^5.1.0",
+    "@react-navigation/native": "^5.0.8",
+    "@react-navigation/stack": "^5.1.0",
     "expo": "~36.0.0",
     "firebase": "^7.9.1",
     "react": "~16.9.0",
@@ -16,7 +19,7 @@
     "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.0.tar.gz",
     "react-native-firebase": "^5.6.0",
     "react-native-gesture-handler": "^1.3.0",
-    "react-native-safe-area-context": "0.6.0",
+    "react-native-safe-area-context": "^0.6.4",
     "react-native-screens": "2.0.0-alpha.12",
     "react-native-web": "~0.11.7",
     "react-navigation": "^4.2.1",
