+++ screens/LoadingScreen.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {View, Text, StyleSheet, ActivityIndicator} from 'react-native'
-import * as firebase from 'firebase'
+import firebase from '../firebase'
 
 export default class LoadingScreen extends React.Component {
     componentDidMount() {
